+++ generator/swagen/suite_test.go
@@ -3,11 +3,14 @@ package swagen
 import (
 	"testing"
 
+	"github.com/gopher-fleece/gleece/infrastructure/logger"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 )
 
 func TestSwagenModule(t *testing.T) {
+	// Disable logging to reduce clutter.
+	logger.SetLogLevel(logger.LogLevelNone)
 	RegisterFailHandler(Fail)
 	RunSpecs(t, "Swagen Module Suite")
 }
